backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
python-jose[cryptography]
cryptography
requests
aiofiles
